chore: update env
